pyproject.toml
[project]
name = "resource-radar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.5.1",
    "flask-admin>=1.6.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
